move initializer implementation, some minor changes
